Update styles add token to support darkMode
